Add setup section to CRISP_config.toml template
reference_channel is the reference channel used for drift compensation (DAPI).

alignment_channels is a list of channels used for the edge_alignment step.  Typically this is the reference channel (DAPI).

inpattern and outpattern are format strings for the input and output filenames.  These format strings are python formatted with named parameters for region, position, and channel.  The slice field must be printf formatted (%d, %03d, etc.).