Add option to adjust PSF center in xy
The Gibson-Lanni PSF generator can place the center of the PSF on a single pixel (ny//2, nx//2) or spread that spot over the 4 center pixels. A value of 0 for [padding][cemter_psf_xy] uses a single 1x1 point and results in a sharp, noisy image.  A value of 1 shifts the center by half a pixel in x and y, producing a blurrier, but less noisy result.  The default is 1.0, as the sharper deconvolution was excessively noisy, especially in EDF image outputs.