Implemented new EDF algorithm
The new EDF method adds a tunable gaussian filter controlled by the filtering key.  The low byte (0x00 - 0x0F) controls the strength of blurring, with good values in the 6-8 range.

The 0x10 bit and 0x20 bit add the majority and sub-band filters respectively.  (not recommended)

The new EDF method computes two output images with different wavelet decompositions and computes either the mean, min, or max of the two.  A postprocessing values of 0x0 uses the mean, 0x1 uses the minimum, 0x2 uses the maximum, and 0x3 adapts based on image brightness (min for dark images, max for brightfield).  0x4 clamps output values to within the range of the input image stack.  0x8 assigns the output value to the closest input value.

Defaults are 0x07 for filtering and either 0x0 or 0x3 for postprocessing.